Break state logic and setup router, moving to components folder
